lib: fix ip4_matchnet() to work on Sparc
Sparc has strict alignment rules for the datatypes. Since the contents of an
ipbl file will have any alignment depending on their position in the file this
would cause bus errors. Do a local copy of the data and use that for operations
and depend on the compiler to optimize this away for platforms allowing
arbitrary alignments.

svn path=/trunk/; revision=1587