lib: fix getfile() if user path exists, but no domain path
This will probably not happen in reality, at least not with the usual vpopmail
directory layout, but it can happen in the testcases. In this case the error
code returned by the function can be arbitrary and not the expected ENOENT.
Even a wrong file could have been opened and returned.

svn path=/trunk/; revision=1616