Qsmtpd: refactor getfile()
First: rename the old "Getfile" test to "Getsettings" as that is what it is
testing.

Then factor out the "open one file in a given directory" inside getfile() and
getfileglobal(). This now less optimal in case that both user and domain
directory are longer than 512 bytes and the file searched for is not present in
the user directory as it now does 2 allocations then instead of one, but this
case is rather unusual and whoever has that problem regularly can increase the
temporary buffer size to make it fit.

svn path=/trunk/; revision=1623