Qsmtpd: fix addressing to local IP causing false positive matches
If the IP address given in RCPT TO is shorter than the local IP but matches
the head of it (e.g. local one is 127.0.0.1 and the address contains
127.0.0.10) than the address was matched even if it should not. The matching
for IPv6 addresses is currently overly strict as there are multiple ways to
write down the same IPv6 address correctly but only an exact match is checked
for. I doubt this is of any interest, but when someone really has a problem
then this can easily be fixed.

svn path=/trunk/; revision=1638