Qsmtpd: add userconf_get_buffer() as abstraction around getfile() and loadlistfd()
Because getfile() is specific to the vpopmail backend this shouldn't be used in the filters directly.

Also the "Getfile" test is renamed to "VPop_control" to reflect that.

Some tests for the moment just have a copy of the functions, which will go away and will be replaced by a
test-specific implementation that makes the tests independent of the vpopmail backend.

svn path=/trunk/; revision=1766