Qsmtpd: improve handling of the first string passed to net_writen()
This string must fit in the buffer and contain the whole status code. So just start by copying the whole string in the buffer and then run into the loop that does the line wrapping.

This should hopefully also tell static analyzers that msg[3] will always be initialized before used.

svn path=/trunk/; revision=1858