Qsmtpd: make it legal to call addrparse() with NULL ds
This allows smtp_from() to entirely ignore what is found, but in case multiple mails are injected from the same local sender the domain path does not need to be constructed multiple times.

It makes smtp_from() cleaner, at the cost that a local domainpath will be held in memory longer.

svn path=/trunk/; revision=1889