Qremote: reorder some code in QP recoding
This ensures that idx can't be 0 when the outer while loop is left, which would lead
to an underrun after the loop. It is very likely that this case has been impossible,
but the checks in the code would have allowed such a case to slip through if off would
have been greater than len at the end of the loop (which should also never happen).

Also add some comments about what is going on, that never hurts.

svn path=/trunk/; revision=1891