Qremote: give tls_init() a more sensible API
Until now tls_init() would either return 1 in case off success or would terminate the process. Make it
return 0 in case of success and 1 in case of error, leaving it to the caller to shut down the connection.
Not all cases where the process was terminated are changed yet, but the tls_quit() function is already
gone. This allowed several places where "write(1, ...)" was called directly to be changed to use
write_status_m().

svn path=/trunk/; revision=1926