Qsmtpd: use explicitely named variables for fds to communicate with qmail-queue
Both ends of the pipes are only needed in queue_init(), where qmail-queue is
forked and needs the read ends. All other places only need the write ends of the
pipes, and copied that to a direct variable instead of referencing in the array
anyway. So the pipe array now is local to queue_init() and the explicit
variables are file global, this time with good names about what they contain.

svn path=/trunk/; revision=1948