Qsmtpd: remove the file-global "chunked" variable
This indicated if a BDAT transfer was running, but it was not used this way. The
only interesting information is when writing the Received header and envelope
data if the message was transferred using BDAT, and that is obvious looking at
the calling function.

While at it add some API docs.

svn path=/trunk/; revision=1950