introduce abstraction for opening a directory fd
The reason behind this is that if a directory only has search permissions for
the qmail users (i.e. "--x") than this would be sufficient for opening a file by
path, but one usually can't open() that directory to use the result as anchor
point of openat(). There are ways to do that for some platforms (e.g. O_PATH on
Linux >= 2.6.39). Having the domain and user directories with that permissions
prevents enumerating e.g. all the users of a given domain in case someone gets
access to the qmail or any other unprivileged account on that machine.

svn path=/trunk/; revision=2006