vpopmail backend: if .qmail-default can't be opened don't read from it
The case where one can't open .qmail-default and the case where vpopbounce isn't
given has the same consequences: it cannot be determined if the mail would be
bounced and therefore should be accepted. The recent changes to qmexists() would
cause an attempt to read .qmail-default if the permission to do so is missing,
leading to an error message instead of acceptance of the mail.

svn path=/trunk/; revision=2057