Qsmtpd: fix the config type return value of getsetting*() and most filters
The values of the type argument for getfile() were changed to use an enum with
the relevant values being 1, 2, and 4. Sadly getsetting*() were not changed at
the same time, returning 0, 1, and 2 like they always did. This lead to wrong
or truncated messages which filter matched, but luckily to no further damage.

Use an enum type instead of int for all these flags.

svn path=/trunk/; revision=2058