Qsmtpd: fix the bad bounce code
Reject bad bounces before doing any user filters, this avoids needless network
I/O and calculations just to reject the mail afterwards anyway.

Codewise do not check for an SMTP state to detect if there have been any
recipients, but instead just look at the recipient counter.

Also fix the bug that the struct recip for the first message was removed from
the recipient list, but not freed, so it was actually leaked. Leave it in the
list, as the current recipient is also left in the list. They will all be freed
when the connection is closed or RSET is called. Further fix that the log
message had sender and recipient exchanged and also was printed again for an
already rejected recipient.

svn path=/trunk/; revision=2097