Qsmtpd: remove a fix for an off-by-one that was none
In r2099 a fix for an off-by-one was introduced, that actually cannot happen. The trick here is that alen
already is length+1 so the trailing newline is included, and the if checks for the combined length being
smaller than the buffer. So the additional '.' will make the buffer filled up exactly to it's end, and no
overrun will happen. Add a comment so noone will get fooled by this again.

Math:
The strings are both 127 chars long. alen will be 128, dlen 127. This results in alen+dlen<256, so the buffer
is filled with 127+128+1=256 chars. If any of the strings would be longer the buffer will not be used in
this iteration.

svn path=/trunk/; revision=2120