Qsmtpd: remove 2 parameter names from header
Old gcc versions (e.g. 3.4) complain about this shadowing a global variable. Yes, it would, but the file
where the function is defined, i.e. where the parameters are actually used, does not include commands.h
where the global variable is declared. And in fact the value passed in _is_ actually exactly the global
variable.

svn path=/trunk/; revision=2121