one intermediate step on storing a list of IPs per MX entry
This is less efficient as before, and should not go into production, it is just an intermediate step.
Convert struct ips addr field into a pointer, but for the moment still store the ip itself in the same
struct to avoid useless memory allocations. All consumers assert that the pointer is properly set up, so
nothing should go wrong. Let's see.

svn path=/trunk/; revision=2127