DNS: add a count member to struct ips
This will eventually allow to check the length of the addr member. Assert that is 1 for now at all known
and some new places.

Use freeips() also to free single entries when they are removed from the list to get the additional sanity
checking. This also makes it easier to add freeing of addr in only one place later.

svn path=/trunk/; revision=2142