Qremote: move the call to getrhost() into tryconn()
Originally the call to getrhost() was delayed to avoid the memory allocation for the remote host name
until the connection succeeded. Meanwhile this call was moved forward to give a proper error message if
the first read() or write() fails, so it now comes basically immediately after tryconn() successfully
returns. Now move it inside tryconn() so getrhost() does not need to search for the IP to print, and
while at it make it ready for the new layout of struct ips.

svn path=/trunk/; revision=2155