Qsmtpd: fix more leaking of MAIL FROM:
If MAIL FROM: did not succeed the memory allocated for the from address was not
freed. If the next MAIL FROM: was successful this memory would be leaked. The
same holds for other information from this MAIL FROM: command, e.g. size,
8BITMIME, or MX entries of the remote domain.

Also reset the spacebug flag before each mail from so a new successful delivery
will not be punished for the errors of the previous one.

Most of these cases will only affect deliveries by telnet. Normal SMTP server
implementations would RSET or QUIT on error, and would either always send SIZE
and overwrite a possible stale value or never send it.

svn path=/trunk/; revision=2207