Qsmtpd: check if target domains exist if relaying is permitted
This prevents an authenticated user to send to a non-existing target domain,
i.e. immediately rejects typos in the target domain. Otherwise the message
would sit for 5 days in the queue until it would be bounced.

This is not applied to the from address as you may have a hostname in a local
net sending you e.g. status messages or other such things. For the from address
the "fromdomain" filter serves a similar purpose.

svn path=/trunk/; revision=2231