Qsmtpd: fix memleak if multiple users use rSPF with exp= modifiers
If no authorative SPF entry was found users may override this using their rSPF entries. If a user has a rSPF
list that has an exp= modifier in it's result this string will leak out of cb_spf() as it is recorded in the
global struct xmitstat. It may later be used as an error string by another user with an unrelated rSPF entry,
or if a later recipient uses a rSPF list with an exp= modifier the memory will be lost.

svn path=/trunk/; revision=2259