SSL: fix bad return value on timeout
The change r2277 entirely broke the timeout handling in SSL mode as a negative
timeout would be passed to poll, resulting in a wait for any input, which may
never come. But also before the code was unreliable: if the first e.g. read
would fail because it needed e.g. a rehandshake, then poll() would be called
and may return that the output descriptor is ready. Then the next read would
time out, and the result of the first poll would still be used, signalling that
everything is fine when it is actually not.

svn path=/trunk/; revision=2282