SSL: clean up the interface of the wrapper functions
The first goal is to avoid that the ssl_*() wrappers directly call dieerror(),
let higher levels of the stack handle this. This has the neat side effect that
ssl_timeoutconn() now can fail with a timeout and Qremote would just try the
next server instead of dying.

The other part is that these functions will now either return a negative error
code (from errno.h) or a sane value. That means that ssl_timeoutaccept() and
ssl_timeoutconn() will just return 0 if everything is fine like accept() and
connect() do.

svn path=/trunk/; revision=2284