allow some network errors to be non-fatal
This allows e.g. Qremote to check if reading from the network failed and
properly handling this instead of abusing data_pending() for this. Using
data_pending() has the disadvantage that if the connection is there at the
moment where it is tested, but is closed before any data is read this can't be
detected. Now the poll() before the read() itself will return and tell that the
connection is gone, or the read will return 0 and the caller could also handle
this.

svn path=/trunk/; revision=2286