Qsmtpd: fix crash due to bad truncation
For this to happens some things need to happen at the same time:
-sizeof(long) != sizeof(int) (i.e. 64 bit platform)
-the command has a target state of -1 (i.e. "do not change state")
-this command needs to be accepted and called twice

The target state will be saved and later restored. For this a too short
variable was used, leading to a bad target state afterwards. It was no longer
negative, so after the second command the target state was 0xffffffff, allowing
all commands to happen next.

A sequence that may lead to a crash was:
VRFY
VRFY
MAIL FROM:<>

This happened because MAIL FROM relies on information from HELO or EHLO being
present, which had not happened here.

Found by american fuzzy lop.

svn path=/trunk/; revision=2370