allow configfiles to embed their parent files
This is mainly useful for filter lists like badmailfrom or spfstrict, where
one can have additional rules for specific users, but those do not want to sync
with the domain or global files all the times. Now they can simply drag in all
the contents of those files at will.