Qremote: use OpenSSL 1.0.2 builtin to verify server hostname
The control/tlshosts/ verification code now requires OpenSSL 1.0.2 and will
always produce an error if an attempt is made to use it with a lower version.
This is to avoid accidential bypassing of the certificate verification.

All other SSL code still works with lower OpenSSL versions, so they are still
supported. The minimum required version will be bumped at the end of this year
when OpenSSL 1.0.2 will become the oldest supported upstream branch.