Qsmtpd: fix increasing maximum line length if 'RCPT TO:<...> ... AUTH' is received
The pointer was changed to point to a random stack location, that could be
increased by 26. No idea what that could do, but make sure it doesn't happen
again by making the pointer location const.

Sort-of spotted by Coverity, which warned that something was passed in that is
no array, but was used as one.