tests: fix typo in IP address in all_filters test
The address was broken, resulting in "::" being stored in the MX address. If
IPV4ONLY was set this resulted in this test failing because the address was
outside the valid address range, and it should have failed for IPv6, too, but
that address was not in the list of bad addresses. Use an IPv4 address that
will not change: the A DNS root server.

While at it, make sure the address conversion actually works.