Qsmtpd: fix sequence confusion when check2822 errors are raised at mail end
If the line that caused a check2822 error to be raised is already the
terminating line of the DATA phase, e.g. the "From:" or "Date:" header is
missing and the mail has no separating line to the body, then the terminating
line was not properly detected as such and it was tried to read more data.