tests: fix GCC 11 compiler warnings in NetIO test
The argument to net_readline() are marked as read_write, as the code reads from
that buffer to see if has written a CR to it before. This will only ever happen
if it has actually written anything before. GCC 11 now complains that this may
use the uninitialized test buffer. It can't know that this would never happen.
In any case it seems to be a good idea to poison the buffer before calling the
function to see if all works properly.