Qremote: explicitely set errno to 0 when smtproutes.d value has matched
One could have an entry that does not set the relay host, which is still valid
e.g. when setting just the certificate. Since this would lead to not returning
the MX pointer. When errno has a stale old value this could lead that an error
is raised even if everything is fine. The smtproutes code at the end explicitely
resets errno, so this exit path should do it, too.