remove invalid "pure" attributes
These functions modify a pointer passed to them, so they are not pure. When gcc
then uses the opportunity to optimize based on that pure it can break things,
as observed in the AddrSyntax test on amd64 with recent enough versions of gcc.

Thanks to slyfox for actually finding out what I did wrong.