Qsmtp: fix truncated Received: lines
With recent gcc versions (>=11) the temporary character buffer is never
initialized as gcc seems to expect that noone actually takes a reference on it.
This leads to the Received: line being truncated right after the IP address.
Move the variable to a scope where it stays alive until it's final use.