also terminate on SSL protocol errors when fatal error handling is requested
There are situations where a graceful error handling is not possible, e.g. when
Qsmtpd is in the DATA phase. In these cases certain errors, e.g. timeouts,
cause a immediate program termination. SSL protocol errors where missed to also
trigger this, and instead this could lead to an endless loop where the process
was stuck at 100% CPU usage.